Use compatibilityChecker to detect missing service image as a blocker
